[analyzer] Fix extraction of punned and known scalar SVals
This change addresses punning of scalar types in static analysis, and is
a precursor to handling the following two cases:
   https://bugs.llvm.org/show_bug.cgi?id=39032
   https://bugs.llvm.org/show_bug.cgi?id=44114

This patch was co-authored with Balazs Benics (steakhal), and debug
assistance from Gabor Marton (martong).

Region store is taught to recognize punned concrete integers and extract
the data in an endianess correct way for scalars.

Co-authored-by: Balazs Benics <benicsbalazs@gmail.com>

Differential Revision: https://reviews.llvm.org/D93595